requirements.txt
azure-functions==1.12.0
azure-cosmos==4.3.0
hundredandten==1.4.0
